requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
PyYAML
python-dotenv
tqdm
loguru
streamlit
PysimpleGUI